refactor(settings): extract Updates/Users panels + shared helpers from SettingsApp (S13) (#545)
